requirements.txt
langgraph==1.1.3
langgraph-checkpoint-sqlite==3.0.3
langchain-core==1.2.20
langchain-openai==1.1.11
python-dotenv==1.2.2
pypdf==6.1.1
streamlit==1.49.1
